Volatility_Detector_Binance/requirements.txt
python-binance==1.0.17
pandas==2.0.3
numpy==1.24.3
ta==0.10.2
python-dotenv==1.0.0
